Corrected minor errors in audio and piano

+++ examples/piano.cpp
@@ -2,16 +2,14 @@
 #include <Flaudio.h>
 
 #define _USE_MATH_DEFINES
-#include <cmath>
+#include <math.h>
 #include <iostream>
-
-#include <unistd.h>
 
 using namespace std;
 
 int main(){
   const int w = 400, h = 400;
-  
+
   Winval win(w, h);
 
   const char* chars = "aoeuidhtns";
@@ -19,7 +17,7 @@ int main(){
   const int steps[] = {0, 2, 4, 5, 7, 9, 11, 12, 14, 16};
 
   const int numKeys = 10;
-  
+
   bool isPressed[numKeys];
   int channels[numKeys];
   for(int i =0; i < numKeys; i++){
@@ -28,15 +26,15 @@ int main(){
   }
 
 
-  
-  
+
+
   Flaudio fl;
 
   int samplerate = fl.getSampleRate();
   int16_t *tempsamples= new int16_t[samplerate];
   int16_t maxvol = 1<<11;
 
-  
+
   /*int16_t* test  = new int16_t[samplerate];
   float fruq = 440;
   for(int i= 0;i < samplerate; i++){
@@ -57,7 +55,7 @@ int main(){
     for(int i = 0; i < numKeys; i++){
       if(!isPressed[i] && win.isKeyPressed(charcodes[i])){
 	float freq = 440*powf(2.0f, (steps[i] - 9.0f)/12.0f);
-	int buffsize = samplerate/freq;
+	int buffsize = (int)(samplerate/freq);
 	for(int p =0; p < buffsize; p++){
 	  tempsamples[p] = (int16_t)(maxvol*sin(freq*p/samplerate*2*M_PI));
 	}
@@ -66,7 +64,7 @@ int main(){
 	//cout<<"Started in channel "<<channels[i]<<endl;
       }else if(isPressed[i] && !win.isKeyPressed(charcodes[i])){
 	fl.endSegmentInChannel(channels[i]);
-	
+
 	//cout<<"Ended in channel "<<channels[i]<<endl;
 	channels[i] = -1;
 	isPressed[i] = false;
@@ -82,5 +80,5 @@ int main(){
     }
   }
 
-  return 0; 
+  return 0;
 }
